Optimized ReadIndex (20% faster), and ReadStructMarshal (11 times faster or twice as fast as ReadStruct).
Implemented WriteCompactIndex so we can now more easily re-serialize struct data and verify that serialized data have no mismatches.

Also deprecated the hacky-tracking of LastPosition when reading in data, instead we store the last position on every call to "Record", this approach however requires us to call "ConformRecordPosition" to correct the position in several circumstances.

Implemented more UnrealStreamTests to verify that ReadIndex and WriteIndex work for all use cases, and for each relevant package version change.